[code-infra] Move getReleaseVersion into pnpm.mjs
It is a workspace-manifest query like its neighbours getPublishRegistry and
getPackagesNeedingManualPublish, and composes helpers pnpm.mjs already owns
(findWorkspaceDir, readPackageJson). cmdPublish just calls it now; its tests
move alongside the other pnpm.mjs suites.
